backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Pillow==10.0.1
openpyxl==3.1.2
qrcode==7.4.2
requests==2.31.0
gunicorn==21.2.0
redis==4.6.0
celery==5.3.1
cryptography==41.0.7
Werkzeug==2.3.7
APScheduler==3.10.4
python-magic==0.4.27
bleach==6.1.0